Update refinerycms-carousel engine: -Fix migrations -Update gem dependencies -Mount refinery engine -Still very WIP

+++ db/migrate/20111129234241_remove_cached_slug_from_pages.rb
@@ -1,6 +1,6 @@
 class RemoveCachedSlugFromPages < ActiveRecord::Migration
-  def self.up
-    if ::Refinery::Page.column_names.map(&:to_s).include?('cached_slug')
+  def up
+    if Refinery::Page.column_names.map(&:to_s).include?('cached_slug')
       say_with_time("Removing cached_slug column from ::Refinery::Page table") do
         remove_column ::Refinery::Page.table_name, :cached_slug
       end
@@ -8,10 +8,10 @@ class RemoveCachedSlugFromPages < ActiveRecord::Migration
       say "Nothing done, no cached_slug field found in ::Refinery::Page table"
     end
 
-    ::Refinery::Page.reset_column_information
+    Refinery::Page.reset_column_information
   end
 
-  def self.down
+  def down
     # Don't add this column back, it breaks stuff.
   end
 end
